Fix incorrect user count when using multiple different queries
When you had two usages of useQueryRx with querykey A and querykey B,
subscriptionSubject would fire for both querykeys even if it was just
querykey B that changed. This is because the subject gets reused for all
querykeys. This caused querykey A to be counted as a user for the
subscription of querykey B.

This commit filters out such that subscription changes only count users
for a given querykey if the subscription that has changed was *actually*
for said querykey.